move chan_open_try
